+++ components/elements/Input.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from "react";
 
 import { cn } from "../util";
-import { Label, LabelProps } from "./Label";
+import { Label, LabelProps } from "./label/Label";
 import { Skeleton } from "./skeleton/Skeleton";
 
 type TextFieldTypes = React.InputHTMLAttributes<HTMLInputElement> & {