+++ components/elements/PhoneInput.tsx
@@ -2,7 +2,7 @@ import React, { useState, FC, useRef, useEffect } from "react";
 
 import Countries from "../countries";
 import { cn } from "../util";
-import { Label, LabelProps } from "./Label";
+import { Label, LabelProps } from "./label/Label";
 import { Select } from "./Select";
 
 type PhoneInputTypes = {