+++ components/elements/Select.tsx
@@ -5,8 +5,8 @@ import CreatableSelect from "react-select/creatable";
 import clsx from "clsx";
 
 import { cn } from "../util";
-import { Label, LabelProps } from "./Label";
-import { Skeleton } from "./skeleton/Skeleton";
+import { Label, LabelProps } from "./label";
+import { Skeleton } from "./skeleton";
 
 type ControlTypes = {
   cx: any;